修改数据库后重新生曾的pojo和mapper

+++ zj22-gudao-server/gudao-server-web/src/main/java/me/zj22/gudao/server/web/service/OrderService.java
@@ -16,7 +16,7 @@ public interface OrderService {
     OrderDTO create(OrderDTO orderDTO);
 
     /**查询单个订单*/
-    OrderDTO findOne(Integer orderId);
+    OrderDTO findOne(Long orderId);
 
     /**查询订单列表,分页  里面的keyWord也填充为openId*/
     Page<OrderDTO> findAllListOrder(Page<OrderDTO> page, String openId);
